move reaction to message handler
